updated roya-workflow to have continuations and nested workflow as first-class features

+++ workflow-orchestrator/src/test/java/com/workflow/WorkflowTest.java
@@ -1,7 +1,6 @@
 package com.akilisha.oss.roya.workflow;
 
-import com.akilisha.oss.roya.workflow.core.NodeOutput;
-import com.akilisha.oss.roya.workflow.core.Workflow;
+import com.akilisha.oss.roya.workflow.core.*;
 import com.akilisha.oss.roya.workflow.edges.Edge;
 import com.akilisha.oss.roya.workflow.execution.WorkflowExecutor;
 import com.akilisha.oss.roya.workflow.execution.WorkflowResult;
@@ -10,19 +9,18 @@ import org.junit.jupiter.api.Test;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Example test demonstrating workflow testing patterns
  */
 public class WorkflowTest {
-
+    
     @Test
     void testSimpleSequentialWorkflow() {
         // Given: A simple two-node workflow
         Workflow workflow = Workflow.create()
-            .trigger("start", input ->
+            .trigger("start", input -> 
                 CompletableFuture.completedFuture(
                     NodeOutput.success("value", 10)
                 ))
@@ -34,17 +32,17 @@ public class WorkflowTest {
             })
             .edge("start", "double")
             .build();
-
+        
         // When: Execute workflow
         WorkflowResult result = new WorkflowExecutor(workflow)
             .executeFrom("start", Map.of())
             .join();
-
+        
         // Then: Verify success and result
         assertTrue(result.isSuccess());
-        assertEquals(20, (Integer) result.context().get("result"));
+        assertEquals(20, result.context().get("result"));
     }
-
+    
     @Test
     void testConditionalBranching() {
         // Given: Workflow with conditional branching
@@ -55,21 +53,21 @@ public class WorkflowTest {
                 CompletableFuture.completedFuture(NodeOutput.success("path", "A")))
             .action("pathB", input ->
                 CompletableFuture.completedFuture(NodeOutput.success("path", "B")))
-
+            
             .edge("start", "pathA", Edge.when(ctx -> ctx.<Integer>get("age") >= 18))
             .edge("start", "pathB", Edge.when(ctx -> ctx.<Integer>get("age") < 18))
             .build();
-
+        
         // When: Execute with age >= 18
         WorkflowResult result = new WorkflowExecutor(workflow)
             .executeFrom("start", Map.of("age", 25))
             .join();
-
+        
         // Then: Should take path A
         assertTrue(result.isSuccess());
         assertEquals("A", result.context().get("path"));
     }
-
+    
     @Test
     void testParallelExecution() {
         // Given: Workflow with parallel branches
@@ -84,22 +82,22 @@ public class WorkflowTest {
                 input.context().set("completedB", true);
                 return CompletableFuture.completedFuture(NodeOutput.success());
             })
-
+            
             .edge("start", "taskA", Edge.parallel())
             .edge("start", "taskB", Edge.parallel())
             .build();
-
+        
         // When: Execute
         WorkflowResult result = new WorkflowExecutor(workflow)
             .executeFrom("start", Map.of())
             .join();
-
+        
         // Then: Both tasks should complete
         assertTrue(result.isSuccess());
         assertTrue((Boolean) result.context().get("completedA"));
         assertTrue((Boolean) result.context().get("completedB"));
     }
-
+    
     @Test
     void testNodeFailure() {
         // Given: Workflow with failing node
@@ -108,21 +106,21 @@ public class WorkflowTest {
                 CompletableFuture.completedFuture(NodeOutput.success()))
             .action("fail", input ->
                 CompletableFuture.completedFuture(NodeOutput.failure("Intentional failure")))
-
+            
             .edge("start", "fail")
             .build();
-
+        
         // When: Execute
         WorkflowResult result = new WorkflowExecutor(workflow)
             .executeFrom("start", Map.of())
             .join();
-
+        
         // Then: Workflow should fail
         assertTrue(result.isFailure());
         assertTrue(result.finalOutput().error().isPresent());
         assertTrue(result.finalOutput().error().get().contains("Intentional failure"));
     }
-
+    
     @Test
     void testExecutionTrace() {
         // Given: Multi-node workflow
@@ -130,23 +128,23 @@ public class WorkflowTest {
             .trigger("start", input -> CompletableFuture.completedFuture(NodeOutput.success()))
             .action("step1", input -> CompletableFuture.completedFuture(NodeOutput.success()))
             .action("step2", input -> CompletableFuture.completedFuture(NodeOutput.success()))
-
+            
             .edge("start", "step1")
             .edge("step1", "step2")
             .build();
-
+        
         // When: Execute
         WorkflowResult result = new WorkflowExecutor(workflow)
             .executeFrom("start", Map.of())
             .join();
-
+        
         // Then: All nodes should be in trace
         assertEquals(3, result.getTrace().size());
         assertEquals("start", result.getTrace().get(0).nodeId());
         assertEquals("step1", result.getTrace().get(1).nodeId());
         assertEquals("step2", result.getTrace().get(2).nodeId());
     }
-
+    
     @Test
     void testContextDataPropagation() {
         // Given: Workflow that transforms data
@@ -159,15 +157,15 @@ public class WorkflowTest {
                     NodeOutput.success("greeting", "Hello, " + name + "!")
                 );
             })
-
+            
             .edge("start", "greet")
             .build();
-
+        
         // When: Execute
         WorkflowResult result = new WorkflowExecutor(workflow)
             .executeFrom("start", Map.of())
             .join();
-
+        
         // Then: Data should propagate through context
         assertEquals("Alice", result.context().get("name"));
         assertEquals("Hello, Alice!", result.context().get("greeting"));
